# perf: specialize is_neutral_element for each curve