Create `is_neutral_element` implementations for Montgomery, Short Weiestrass and Edwards curves.

## Type of change

- [x] Optimization

## Checklist
- [x] This change is an Optimization
  - [x] Benchmarks added/run

## Benchmark results

TL;DR: 25-40% speedup.

```
MSM benchmarks with size 2/Naive                                                                                                                                                                                                              
                        time:   [1.0780 ms 1.0783 ms 1.0785 ms]                                                                                                                                                                               
                        change: [-26.694% -26.659% -26.623%] (p = 0.00 < 0.05)                                                        
MSM benchmarks with size 2/Sequential Pippenger/1                                                                                                                                                                                             
                        time:   [1.0874 ms 1.0877 ms 1.0879 ms]                                                                                                                                                                               
                        change: [-37.072% -36.949% -36.828%] (p = 0.00 < 0.05)                                                                                                                                                              
MSM benchmarks with size 2/Parallel Pippenger/1                                                                                                                                                                                               
                        time:   [4.1959 ms 4.2006 ms 4.2051 ms]                                                                                                                                                                               
                        change: [-27.033% -26.910% -26.789%] (p = 0.00 < 0.05)                                                               
MSM benchmarks with size 2/Sequential Pippenger/2                                                                                                                                                                                             
                        time:   [1.0839 ms 1.0844 ms 1.0849 ms]                                                                                                                                                                               
                        change: [-38.057% -37.997% -37.936%] (p = 0.00 < 0.05)                                                                                                                                                      
MSM benchmarks with size 2/Parallel Pippenger/2                                                                                                                                                                                               
                        time:   [2.6903 ms 2.6937 ms 2.6970 ms]                                                                                                                                                                               
                        change: [-25.115% -24.971% -24.819%] (p = 0.00 < 0.05)
MSM benchmarks with size 2/Sequential Pippenger/4                                                                                                                                                                                             
                        time:   [1.4155 ms 1.4159 ms 1.4163 ms]                                                        
                        change: [-39.365% -39.332% -39.298%] (p = 0.00 < 0.05)
MSM benchmarks with size 2/Parallel Pippenger/4
                        time:   [1.6254 ms 1.6279 ms 1.6303 ms]
                        change: [-25.903% -25.659% -25.361%] (p = 0.00 < 0.05)                                   
MSM benchmarks with size 2/Sequential Pippenger/8                                                                                                                                                                                             
                        time:   [7.5383 ms 7.5410 ms 7.5437 ms]
                        change: [-42.754% -42.659% -42.561%] (p = 0.00 < 0.05)
MSM benchmarks with size 2/Parallel Pippenger/8
                        time:   [1.5617 ms 1.5659 ms 1.5703 ms]
                        change: [-32.997% -32.658% -32.329%] (p = 0.00 < 0.05)
MSM benchmarks with size 2/Sequential Pippenger/12                                                                                                                                                                                   [61/1958]
                        time:   [78.514 ms 78.539 ms 78.565 ms]
                        change: [-42.778% -42.671% -42.565%] (p = 0.00 < 0.05)
MSM benchmarks with size 2/Parallel Pippenger/12
                        time:   [6.9643 ms 6.9937 ms 7.0235 ms]
                        change: [-41.578% -41.157% -40.728%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Naive
                        time:   [2.1828 ms 2.1847 ms 2.1864 ms]
                        change: [-25.940% -25.842% -25.743%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Sequential Pippenger/1
                        time:   [1.3428 ms 1.3438 ms 1.3448 ms]
                        change: [-37.176% -37.087% -37.001%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Parallel Pippenger/1
                        time:   [5.0040 ms 5.0086 ms 5.0133 ms]
                        change: [-25.249% -25.141% -25.027%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Sequential Pippenger/2
                        time:   [1.3330 ms 1.3361 ms 1.3390 ms]
                        change: [-37.147% -37.013% -36.880%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Parallel Pippenger/2
                        time:   [2.8434 ms 2.8471 ms 2.8509 ms]
                        change: [-24.671% -24.533% -24.392%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Sequential Pippenger/4
                        time:   [1.6949 ms 1.6974 ms 1.7007 ms]
                        change: [-37.099% -36.952% -36.755%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Parallel Pippenger/4
                        time:   [1.6528 ms 1.6551 ms 1.6574 ms]
                        change: [-25.658% -25.445% -25.240%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Sequential Pippenger/8
                        time:   [8.3688 ms 8.3824 ms 8.3960 ms]
                        change: [-40.983% -40.814% -40.625%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Parallel Pippenger/8
                        time:   [1.6157 ms 1.6198 ms 1.6240 ms]
                        change: [-33.195% -32.738% -32.332%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Sequential Pippenger/12
                        time:   [85.617 ms 85.652 ms 85.688 ms]
                        change: [-41.297% -41.239% -41.181%] (p = 0.00 < 0.05)
MSM benchmarks with size 4/Parallel Pippenger/12
                        time:   [7.1005 ms 7.1322 ms 7.1661 ms]
                        change: [-43.183% -42.756% -42.319%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Naive
                        time:   [4.2555 ms 4.2568 ms 4.2581 ms]
                        change: [-27.461% -27.414% -27.367%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Sequential Pippenger/1
                        time:   [1.7517 ms 1.7521 ms 1.7526 ms]
                        change: [-36.385% -36.355% -36.324%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Parallel Pippenger/1
                        time:   [5.4060 ms 5.4105 ms 5.4150 ms]
                        change: [-24.699% -24.599% -24.495%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Sequential Pippenger/2
                        time:   [1.7651 ms 1.7696 ms 1.7743 ms]
                        change: [-36.094% -35.900% -35.706%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Parallel Pippenger/2
                        time:   [2.9347 ms 2.9393 ms 2.9439 ms]
                        change: [-24.370% -24.208% -24.046%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Sequential Pippenger/4 
                        time:   [2.0454 ms 2.0472 ms 2.0489 ms]
                        change: [-37.077% -37.000% -36.922%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Parallel Pippenger/4                                                                        
                        time:   [1.7121 ms 1.7148 ms 1.7175 ms]
                        change: [-25.538% -25.332% -25.143%] (p = 0.00 < 0.05)                          
MSM benchmarks with size 8/Sequential Pippenger/8                                                                      
                        time:   [9.3000 ms 9.3031 ms 9.3061 ms]               
                        change: [-39.749% -39.718% -39.689%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Parallel Pippenger/8
                        time:   [1.7033 ms 1.7076 ms 1.7118 ms]
                        change: [-32.406% -32.116% -31.835%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Sequential Pippenger/12
                        time:   [92.232 ms 92.411 ms 92.588 ms]
                        change: [-40.511% -40.384% -40.277%] (p = 0.00 < 0.05)
MSM benchmarks with size 8/Parallel Pippenger/12
                        time:   [7.3891 ms 7.4204 ms 7.4534 ms]
                        change: [-43.906% -43.556% -43.188%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Naive
                        time:   [8.6346 ms 8.6440 ms 8.6528 ms]
                        change: [-25.961% -25.876% -25.794%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Sequential Pippenger/1
                        time:   [2.6303 ms 2.6355 ms 2.6406 ms]
                        change: [-33.091% -32.941% -32.815%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Parallel Pippenger/1
                        time:   [5.4915 ms 5.4962 ms 5.5009 ms]
                        change: [-24.514% -24.417% -24.322%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Sequential Pippenger/2
                        time:   [2.6571 ms 2.6580 ms 2.6589 ms]
                        change: [-33.451% -33.342% -33.234%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Parallel Pippenger/2
                        time:   [2.9966 ms 3.0002 ms 3.0037 ms]
                        change: [-24.713% -24.569% -24.422%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Sequential Pippenger/4                                                                                                                                                                                                           
                        time:   [2.5965 ms 2.5971 ms 2.5977 ms]
                        change: [-36.997% -36.967% -36.936%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Parallel Pippenger/4
                        time:   [1.7541 ms 1.7569 ms 1.7597 ms]
                        change: [-25.341% -25.131% -24.914%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Sequential Pippenger/8
                        time:   [10.167 ms 10.170 ms 10.173 ms]
                        change: [-38.026% -38.001% -37.977%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Parallel Pippenger/8
                        time:   [1.7838 ms 1.7880 ms 1.7922 ms]
                        change: [-31.708% -31.376% -31.047%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Sequential Pippenger/12
                        time:   [96.583 ms 96.681 ms 96.786 ms]
                        change: [-38.845% -38.779% -38.700%] (p = 0.00 < 0.05)
MSM benchmarks with size 16/Parallel Pippenger/12
                        time:   [7.7550 ms 7.7861 ms 7.8191 ms]
                        change: [-42.985% -42.660% -42.340%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Naive
                        time:   [17.405 ms 17.429 ms 17.453 ms]
                        change: [-25.610% -25.504% -25.400%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Sequential Pippenger/1
                        time:   [4.3730 ms 4.3772 ms 4.3814 ms]
                        change: [-32.576% -32.509% -32.432%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Parallel Pippenger/1
                        time:   [5.6549 ms 5.6600 ms 5.6652 ms]
                        change: [-24.610% -24.516% -24.412%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Sequential Pippenger/2
                        time:   [4.3674 ms 4.3687 ms 4.3701 ms]
                        change: [-32.782% -32.753% -32.722%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Parallel Pippenger/2
                        time:   [3.0984 ms 3.1025 ms 3.1067 ms]
                        change: [-25.151% -25.004% -24.854%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Sequential Pippenger/4
                        time:   [3.6683 ms 3.6724 ms 3.6768 ms]
                        change: [-35.627% -35.557% -35.473%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM benchmarks with size 32/Parallel Pippenger/4
                        time:   [1.8490 ms 1.8528 ms 1.8566 ms]
                        change: [-25.358% -25.129% -24.909%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Sequential Pippenger/8                                                                     
                        time:   [10.870 ms 10.873 ms 10.877 ms]               
                        change: [-38.246% -38.114% -37.981%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Parallel Pippenger/8                                                                       
                        time:   [1.8569 ms 1.8615 ms 1.8660 ms]               
                        change: [-32.283% -31.900% -31.529%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Sequential Pippenger/12                                                                    
                        time:   [98.547 ms 98.576 ms 98.606 ms]               
                        change: [-38.620% -38.597% -38.575%] (p = 0.00 < 0.05)
MSM benchmarks with size 32/Parallel Pippenger/12
                        time:   [7.9862 ms 8.0137 ms 8.0429 ms]
                        change: [-42.666% -42.389% -42.083%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Naive                                                                                      
                        time:   [34.730 ms 34.793 ms 34.854 ms]               
                        change: [-23.933% -23.791% -23.656%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Sequential Pippenger/1                                                                     
                        time:   [7.7311 ms 7.7441 ms 7.7572 ms]               
                        change: [-31.181% -31.058% -30.934%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Parallel Pippenger/1                                                                       
                        time:   [5.9657 ms 5.9722 ms 5.9786 ms]               
                        change: [-24.932% -24.821% -24.711%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Sequential Pippenger/2
                        time:   [7.7267 ms 7.7414 ms 7.7562 ms]
                        change: [-31.772% -31.642% -31.521%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Parallel Pippenger/2                                                                       
                        time:   [3.3291 ms 3.3327 ms 3.3364 ms]
                        change: [-25.439% -25.305% -25.178%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Sequential Pippenger/4                                                                     
                        time:   [5.8627 ms 5.8689 ms 5.8752 ms]
                        change: [-33.947% -33.877% -33.808%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Parallel Pippenger/4                                                                                                                                                                                      
                        time:   [2.0062 ms 2.0096 ms 2.0130 ms]                        
                        change: [-25.733% -25.523% -25.317%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Sequential Pippenger/8
                        time:   [12.237 ms 12.242 ms 12.247 ms]
                        change: [-37.278% -37.249% -37.220%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Parallel Pippenger/8
                        time:   [1.9649 ms 1.9695 ms 1.9743 ms]
                        change: [-32.085% -31.742% -31.405%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Sequential Pippenger/12
                        time:   [101.85 ms 102.06 ms 102.27 ms]
                        change: [-38.773% -38.626% -38.469%] (p = 0.00 < 0.05)
MSM benchmarks with size 64/Parallel Pippenger/12
                        time:   [8.1429 ms 8.1673 ms 8.1928 ms]
                        change: [-42.613% -42.349% -42.081%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Naive
                        time:   [68.230 ms 68.248 ms 68.267 ms]
                        change: [-26.772% -26.739% -26.706%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Sequential Pippenger/1
                        time:   [14.364 ms 14.368 ms 14.372 ms]
                        change: [-32.516% -32.481% -32.444%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Parallel Pippenger/1
                        time:   [6.5578 ms 6.5635 ms 6.5691 ms]
                        change: [-25.654% -25.551% -25.442%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Sequential Pippenger/2
                        time:   [14.655 ms 14.660 ms 14.666 ms]
                        change: [-30.975% -30.944% -30.912%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Parallel Pippenger/2
                        time:   [3.8036 ms 3.8083 ms 3.8131 ms]
                        change: [-26.272% -26.143% -26.011%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Sequential Pippenger/4
                        time:   [9.9391 ms 9.9405 ms 9.9418 ms]
                        change: [-33.621% -33.515% -33.405%] (p = 0.00 < 0.05)
MSM benchmarks with size 128/Parallel Pippenger/4                                                                                                                                                                                   
                        time:   [2.2862 ms 2.2893 ms 2.2925 ms]                                                        
                        change: [-27.499% -27.320% -27.140%] (p = 0.00 < 0.05)                                         
MSM benchmarks with size 128/Sequential Pippenger/8
                        time:   [14.684 ms 14.699 ms 14.713 ms]                                                        
                        change: [-37.377% -37.308% -37.236%] (p = 0.00 < 0.05)                                         
MSM benchmarks with size 128/Parallel Pippenger/8                                                                      
                        time:   [2.1579 ms 2.1633 ms 2.1689 ms]                                                        
                        change: [-32.539% -32.222% -31.909%] (p = 0.00 < 0.05)                                         
MSM benchmarks with size 128/Sequential Pippenger/12 
                        time:   [104.86 ms 104.89 ms 104.92 ms]                                                        
                        change: [-38.422% -38.318% -38.212%] (p = 0.00 < 0.05)                                         
…